database: Rename crate name from db to database.
