+++ .gitignore
@@ -5,4 +5,4 @@ risc0-guests/verifier/guest/Cargo.lock
 target/
 .env
 configs
-database
+./database
