+++ core/src/database/common.rs
