+++ core/src/database/verifier.rs
